requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.3.0
mercadobitcoin>=0.1.0
